changed range of crusher set operation to 0.0 to 1.0

+++ LEAF/Inc/leaf-crusher.h
@@ -42,7 +42,7 @@ void    tCrusher_free    (tCrusher* const);
 float   tCrusher_tick    (tCrusher* const, float input);
     
 // 0-7
-void    tCrusher_setOperation (tCrusher* const, uint32_t op);
+void    tCrusher_setOperation (tCrusher* const, float op);
  
 // 0.0 - 1.0
 void    tCrusher_setQuality (tCrusher* const, float val);
